+++ Android/UdacityExercises/Lesson09b-ToDo-List-AAC/T09b.10-Exercise-AddViewModelToAddTaskActivity/app/src/main/java/com/example/android/todolist/AddTaskViewModelFactory.java
@@ -1,17 +1,23 @@
 package com.example.android.todolist;
 
-// TODO (1) Make this class extend ViewModel ViewModelProvider.NewInstanceFactory
-public class AddTaskViewModelFactory {
+import android.arch.lifecycle.ViewModel;
+import android.arch.lifecycle.ViewModelProvider;
 
-    // TODO (2) Add two member variables. One for the database and one for the taskId
+import com.example.android.todolist.database.AppDatabase;
 
-    // TODO (3) Initialize the member variables in the constructor with the parameters received
+public class AddTaskViewModelFactory extends ViewModelProvider.NewInstanceFactory {
 
-    // TODO (4) Uncomment the following method
-    // Note: This can be reused with minor modifications
-    /*@Override
+    private final AppDatabase mDb;
+    private final int mTaskId;
+
+    public AddTaskViewModelFactory(AppDatabase database, int taskId) {
+        mDb = database;
+        mTaskId = taskId;
+    }
+
+    @Override
     public <T extends ViewModel> T create(Class<T> modelClass) {
         //noinspection unchecked
         return (T) new AddTaskViewModel(mDb, mTaskId);
-    }*/
+    }
 }