+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### 19.09.2018
 
+* [Android] ToDo List: Adding the ViewModel to AddTaskActivity
 * [Android] ToDo List: Adding the ViewModel
 
 #### 18.09.2018